+++ doc/changes.rst
@@ -6,6 +6,8 @@ Changes
 v0.5.0
 ------
   * Create the :mod:`~sksparse.amd` module, which provides the AMD ordering method.
+  * Create the :mod:`~sksparse.camd` module, which provides the constrained AMD
+    ordering method.
   * Create the :mod:`~sksparse.colamd` module, which provides the COLAMD
     ordering method.